+++ finagle-mysql/src/test/scala/com/twitter/finagle/mysql/unit/RollbackFactoryTest.scala
@@ -2,7 +2,7 @@ package com.twitter.finagle.mysql
 
 import com.twitter.conversions.DurationOps._
 import com.twitter.finagle.stats.NullStatsReceiver
-import com.twitter.finagle.{ClientConnection, Service, ServiceFactory}
+import com.twitter.finagle.{ChannelClosedException, ClientConnection, Service, ServiceFactory}
 import com.twitter.util.{Await, Awaitable, Future, Time}
 import org.scalatest.FunSuite
 
@@ -83,6 +83,36 @@ class RollbackFactoryTest extends FunSuite {
         }
       }
       def apply(c: ClientConnection) = Future.value(svc)
+      def close(deadline: Time): Future[Unit] = svc.close(deadline)
+    }
+
+    val rollbackClient = new RollbackFactory(client, NullStatsReceiver)
+
+    await(rollbackClient().flatMap { svc =>
+      svc(QueryRequest("1")).ensure { svc.close() }
+    })
+
+    assert(requests == Seq(QueryRequest("1"), PoisonConnectionRequest))
+    assert(closeCalled)
+  }
+
+  test("poison request is sent when rollback fails with ChannelClosedException") {
+    var requests: Seq[Request] = Seq.empty
+    var closeCalled = false
+    val client: ServiceFactory[Request, Result] = new ServiceFactory[Request, Result] {
+      private[this] val svc: Service[Request, Result] = new Service[Request, Result] {
+        def apply(req: Request): Future[EOF] = req match {
+          case QueryRequest("ROLLBACK") => Future.exception(new ChannelClosedException())
+          case _ =>
+            requests = requests :+ req
+            Future.value(EOF(0, ServerStatus(0)))
+        }
+        override def close(when: Time): Future[Unit] = {
+          closeCalled = true
+          Future.Done
+        }
+      }
+      def apply(c: ClientConnection): Future[Service[Request, Result]] = Future.value(svc)
       def close(deadline: Time): Future[Unit] = svc.close(deadline)
     }